actualizacion y eliminacion de actividad
